+++ src/datasources/results/components/variable-editors/ResultsVariableQueryEditor.test.tsx
@@ -9,6 +9,7 @@ import { ResultsVariableProperties } from 'datasources/results/types/QueryResult
 import { QueryStepsDataSource } from 'datasources/results/query-handlers/query-steps/QueryStepsDataSource';
 import { ResultsDataSourceBase } from 'datasources/results/ResultsDataSourceBase';
 import React from 'react';
+import userEvent from '@testing-library/user-event';
 
 const fakeWorkspaces: Workspace[] = [
   {
@@ -130,10 +131,63 @@ describe('ResultsVariableQueryEditor', () => {
       queryByResults = screen.getByText('Query by results properties');
       queryBySteps = screen.getByText('Query by steps properties');
 
-      expect(queryByResults).toBeInTheDocument();
-      expect(queryBySteps).toBeInTheDocument();
+    expect(queryByResults).toBeInTheDocument();
+    expect(queryBySteps).toBeInTheDocument();
+  });
+
+  describe('Take input field', () => {
+    it('should render take input field with 1000 as value by default', () => {
+      renderEditor({
+        refId: '',
+        queryType: QueryType.Steps,
+        queryByResults: 'resultsQuery',
+        queryBySteps: '',
+      } as unknown as ResultsQuery);
+
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+      expect(takeInput).toBeInTheDocument();
+      expect(takeInput).toHaveValue(1000);
+    });
+
+    it('should render with existing take when take is already set', () => {
+      renderEditor({
+        refId: '',
+        queryType: QueryType.Steps,
+        queryByResults: 'resultsQuery',
+        queryBySteps: '',
+        take: 2000,
+      } as unknown as ResultsQuery);
+
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+      expect(takeInput).toHaveValue(2000);
+    });
+
+    it('should only allows numbers in Take field', async () => {
+      renderEditor({
+        refId: '',
+        queryType: QueryType.Steps,
+        queryByResults: 'resultsQuery',
+        queryBySteps: '',
+        take: 2000,
+      } as unknown as ResultsQuery);
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+
+      // User tries to enter a non-numeric value
+      await userEvent.clear(takeInput);
+      await userEvent.type(takeInput, 'abc');
+      await waitFor(() => {
+        expect(takeInput).toHaveValue(null);
+      });
+
+      // User enters a valid numeric value
+      await userEvent.clear(takeInput);
+      await userEvent.type(takeInput, '500');
+      await waitFor(() => {
+        expect(takeInput).toHaveValue(500);
+      });
     });
   });
+});
 
   describe('dependencies', () => {
     it('should load part numbers and workspaces on mount', async () => {